iot-agent/requirements.txt
python-pkcs11
requests

